# Faster strongly typed features for server FeatureCollection