--- a/src/Servers/Kestrel/Core/src/Internal/Http/HttpProtocol.FeatureCollection.cs
+++ b/src/Servers/Kestrel/Core/src/Internal/Http/HttpProtocol.FeatureCollection.cs
@@ -16,23 +16,11 @@ using Microsoft.Net.Http.Headers;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
 {
-    internal partial class HttpProtocol : IHttpRequestFeature,
-                                          IHttpResponseFeature,
-                                          IHttpResponseBodyFeature,
-                                          IRouteValuesFeature,
-                                          IEndpointFeature,
-                                          IHttpRequestIdentifierFeature,
-                                          IHttpRequestTrailersFeature,
-                                          IHttpUpgradeFeature,
-                                          IRequestBodyPipeFeature,
-                                          IHttpConnectionFeature,
-                                          IHttpRequestLifetimeFeature,
-                                          IHttpBodyControlFeature,
-                                          IHttpMaxRequestBodySizeFeature,
-                                          IHttpRequestBodyDetectionFeature
+    internal partial class HttpProtocol
     {
         // NOTE: When feature interfaces are added to or removed from this HttpProtocol class implementation,
-        // then the list of `implementedFeatures` in the generated code project MUST also be updated.
+        // then the list of `implementedFeatures` in the generated code project MUST also be updated first
+        // and the code generator re-reun, which will change the interface list.
         // See also: tools/CodeGenerator/HttpProtocolFeatureCollection.cs
 
         string IHttpRequestFeature.Protocol