… src/Servers/Kestrel/Core/src/Internal/Http3/Http3Stream.cs
… src/Servers/Kestrel/perf/Microbenchmarks/HttpProtocolFeatureCollection.cs
… src/Servers/Kestrel/shared/TransportConnection.FeatureCollection.cs
… src/Servers/Kestrel/shared/TransportConnection.Generated.cs
… src/Servers/Kestrel/shared/TransportMultiplexedConnection.FeatureCollection.cs
… src/Servers/Kestrel/shared/TransportMultiplexedConnection.Generated.cs
… src/Servers/Kestrel/tools/CodeGenerator/FeatureCollectionGenerator.cs
… src/Servers/Kestrel/tools/CodeGenerator/HttpProtocolFeatureCollection.cs
… src/Servers/Kestrel/tools/CodeGenerator/TransportConnectionFeatureCollection.cs
… src/Servers/Kestrel/tools/CodeGenerator/TransportMultiplexedConnectionFeatureCollection.cs