Improves access by +50% for commonly used features
|                                         Method |      Mean |          Op/s |      Diff |
|----------------------------------------------- |----------:|--------------:|----------:|
|                Get&lt;IHttpRequestFeature&gt;* |  8.507 ns | 117,554,189.6 |    +50.0% |
|               Get&lt;IHttpResponseFeature&gt;* |  9.034 ns | 110,689,963.7 |         - |
|           Get&lt;IHttpResponseBodyFeature&gt;* |  9.466 ns | 105,636,431.7 |    +58.7% |
|                Get&lt;IRouteValuesFeature&gt;* | 10.007 ns |  99,927,927.4 |    +50.0% |
|                   Get&lt;IEndpointFeature&gt;* | 10.564 ns |  94,656,794.2 |    +44.7% |

Resolves https://github.com/dotnet/aspnetcore/issues/31321

Performance may be able to be improved depending on https://github.com/dotnet/runtime/issues/49614 by having the casts dropped since they are strongly typed.


Also drops the `CastHelpers.ChkCastAny` that it currently makes for each return

![image](https://user-images.githubusercontent.com/1142958/112739023-25dffa00-8f60-11eb-9b6b-854ca9a31edd.png)