@benaadams 2021-03-28
Still calling `ChkCastAny` but that would be now for `runtime` to address