@benaadams 2021-03-28
Can you link the issue? Even in the generic method?

Is https://github.com/dotnet/runtime/issues/49614#issuecomment-808824607 currently