@benaadams 2021-03-28
K now it gets rid of the per item returned (update Jit issue with this https://github.com/dotnet/runtime/issues/49614#issuecomment-808841044)

Before

![image](https://user-images.githubusercontent.com/1142958/112742272-05766680-8f85-11eb-8fdf-3e5c06b75067.png)

After

![image](https://user-images.githubusercontent.com/1142958/112742276-158e4600-8f85-11eb-96f0-aa72fe1f08b9.png)