@benaadams 2021-03-28
Does the benchmark command still work? Can't remember what it is to trigger it /cc @sebastienros