@benaadams 2021-03-28
`IHttpRequestFeature` is the first feature and `IHttpResetFeature` is the 28th feature

|                            Method | branch |      Mean |          Op/s |  delta |
|---------------------------------- |------- |----------:|--------------:|-------:|
| GetViaGeneric_IHttpRequestFeature |   main | 12.919 ns |  77,404,380.4 |      - |
| GetViaGeneric_IHttpRequestFeature |     PR |  8.022 ns | 124,661,077.5 | +61.0% |
|   GetViaGeneric_IHttpResetFeature |   main | 31.514 ns |  31,731,541.5 |      - |
|   GetViaGeneric_IHttpResetFeature |     PR | 28.693 ns |  34,851,283.6 |  +9.8% |
|                                   |        |           |               |        |
|              GetViaGeneric_Custom |   main | 29.857 ns |  33,492,664.8 |      - |
|              GetViaGeneric_Custom |     PR | 28.897 ns |  34,606,200.7 |  +3.3% |
|            GetViaGeneric_NotFound |   main | 29.762 ns |  33,600,220.5 |      - |
|            GetViaGeneric_NotFound |     PR | 29.658 ns |  33,718,090.2 |  +0.4% |
|                                   |        |           |               |        |
|  GetViaTypeOf_IHttpRequestFeature |   main |  5.809 ns | 172,139,937.3 |      - |
|  GetViaTypeOf_IHttpRequestFeature |     PR |  5.930 ns | 168,633,511.9 |  -2.0% |
|    GetViaTypeOf_IHttpResetFeature |   main | 52.447 ns |  19,066,700.6 |      - |
|    GetViaTypeOf_IHttpResetFeature |     PR | 52.591 ns |  19,014,592.2 |  -0.3% |
|                                   |        |           |               |        |
|               GetViaTypeOf_Custom |   main | 52.617 ns |  19,005,350.0 |      - |
|               GetViaTypeOf_Custom |     PR | 52.512 ns |  19,043,238.2 |  +2.0% |
|             GetViaTypeOf_NotFound |   main | 52.799 ns |  18,939,827.9 |        |
|             GetViaTypeOf_NotFound |     PR | 52.445 ns |  19,067,773.8 |  +0.7% |