@benaadams 2021-03-28
Was a bit random before (assume for the casting)

|                                  Method |     Mean |         Op/s |
|---------------------------------------- |---------:|-------------:|
|                Get&lt;IHttpRequestFeature&gt; | 12.76 ns | 78,360,108.4 |
|   Get&lt;IHttpRequestBodyDetectionFeature&gt; | 12.47 ns | 80,185,800.7 |
|           Get&lt;IHttpResponseBodyFeature&gt; | 15.02 ns | 66,571,955.0 |
|            Get&lt;IRequestBodyPipeFeature&gt; | 13.38 ns | 74,750,082.6 |
|      Get&lt;IHttpRequestIdentifierFeature&gt; | 13.65 ns | 73,249,366.2 |
|           Get&lt;IServiceProvidersFeature&gt; | 26.15 ns | 38,237,489.5 |
|        Get&lt;IHttpRequestLifetimeFeature&gt; | 15.29 ns | 65,398,349.2 |
|             Get&lt;IHttpConnectionFeature&gt; | 14.72 ns | 67,921,289.6 |
|                Get&lt;IRouteValuesFeature&gt; | 14.98 ns | 66,762,272.6 |
|                Get&lt;IEndpointFeature&gt; | 15.28 ns | 65,435,240.8 |
|         Get&lt;IHttpAuthenticationFeature&gt; | 27.33 ns | 36,590,896.3 |
|        Get&lt;IHttpRequestTrailersFeature&gt; | 15.97 ns | 62,618,354.1 |
|                      Get&lt;IQueryFeature&gt; | 30.02 ns | 33,309,260.5 |
|                       Get&lt;IFormFeature&gt; | 27.21 ns | 36,757,280.6 |
|                Get&lt;IHttpUpgradeFeature&gt; | 17.11 ns | 58,432,362.5 |
|              Get&lt;IHttp2StreamIdFeature&gt; | 28.53 ns | 35,044,905.9 |
|       Get&lt;IHttpResponseTrailersFeature&gt; | 32.29 ns | 30,972,086.8 |
|            Get&lt;IResponseCookiesFeature&gt; | 28.22 ns | 35,442,007.1 |
|                      Get&lt;IItemsFeature&gt; | 29.91 ns | 33,436,436.6 |
|              Get&lt;ITlsConnectionFeature&gt; | 30.84 ns | 32,425,063.2 |
|              Get&lt;IHttpWebSocketFeature&gt; | 34.29 ns | 29,166,048.5 |
|                    Get&lt;ISessionFeature&gt; | 35.27 ns | 28,350,426.2 |
|     Get&lt;IHttpMaxRequestBodySizeFeature&gt; | 20.53 ns | 48,699,416.4 |
| Get&lt;IHttpMinRequestBodyDataRateFeature&gt; | 23.42 ns | 42,702,194.6 |
|    Get&lt;IHttpMinResponseDataRateFeature&gt; | 23.95 ns | 41,751,489.4 |
|            Get&lt;IHttpBodyControlFeature&gt; | 19.80 ns | 50,514,940.5 |
|                  Get&lt;IHttpResetFeature&gt; | 37.25 ns | 26,845,708.3 |
|               Get&lt;IHttpNotFoundFeature&gt; | 29.45 ns | 33,958,212.4 |